Added cast to fix warning on 32-bit builds.
The call to memcpy in -dataForColumnIndex: was being passed an int (it accepts a size_t, ie: long on some platforms). This generated a warning in Xcode 5 when building (in my instance) for the iPad.

+++ src/FMResultSet.m
@@ -317,7 +317,7 @@
     
     NSMutableData *data = [NSMutableData dataWithLength:(NSUInteger)dataSize];
     
-    memcpy([data mutableBytes], sqlite3_column_blob([_statement statement], columnIdx), dataSize);
+    memcpy([data mutableBytes], sqlite3_column_blob([_statement statement], columnIdx), (size_t)dataSize);
     
     return data;
 }
